pyproject.toml
[tool.poetry]
name = "habits-tracker"
version = "0.1.0"
description = ""
authors = ["Arch4ngel <mishacool4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=4.2.9"
celery = "^5.3.6"
sqlparse = "^0.4.4"
redis = "^5.0.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
drf-yasg = "^1.21.7"
django-celery-beat = "^2.5.0"
rest-framework-simplejwt = "^0.0.2"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^23.5"
setuptools = "^69.1.0"
requests = "^2.31.0"
coverage = "^7.4.1"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
